refactor(sizes): Default sizes in separate folder/file - in preparation for support for multiple sizes
