+++ index.js
@@ -1,48 +1,4 @@
-const unisex = {
-	eu: [35, 36, 36, 37, 38, 38, 39, 39, 40, 41, 42, 43, 44, 45, 46.5],
-	br: [33, 34, 34, 35, 36, 36, 37, 37, 38, 39, 40, 41, 42, 43, 45],
-	cm: [23, 23, 24, 24, 24, 25, 25, 25, 25, 25.7, 26, 26.7, 27.3, 28, 28.6],
-	in: [
-		9,			// 9
-		9.125,	// 9 1/8
-		9.25,		// 9 1/4
-		9.375,	// 9 3/8
-		9.5,		// 9 1/2
-		9.625,	// 9 5/8
-		9.75,		// 9 3/4
-		9.875,	// 9 7/8
-		10,			// 10
-		10.125, // 10 1/8
-		10.25,	// 10 1/4
-		10.5,		// 10 1/2
-		10.75,	// 10 3/4
-		11,			// 11
-		11.25		// 11 1/4
-	]
-};
-
-const defaultSizes = {
-	us: {
-		m: [3.5, 4, 4.5, 5, 5.5, 6, 6.5, 7, 7.5, 8, 8.5, 9, 10.5, 12, 12.5],
-		w: [5, 5.5, 6, 6.5, 7, 7.5, 8, 8.5, 9, 9.5, 10, 11, 12, 13, 14]
-	},
-	ca: {
-		m: [3.5, 4, 4.5, 5, 5.5, 6, 6.5, 7, 7.5, 8, 8.5, 9, 10.5, 12, 12.5],
-		w: [5, 5.5, 6, 6.5, 7, 7.5, 8, 8.5, 9, 9.5, 10, 11, 12, 13, 14]
-	},
-	uk: {
-		m: [3, 3.5, 4, 4.5, 5, 5.5, 6, 6.5, 7, 7.5, 8, 8.5, 10, 11, 12],
-		w: [2.5, 3, 3.5, 4, 4.5, 5, 5.5, 6, 6.5, 7, 8, 9, 10, 11, 12]
-	},
-	au: {
-		m: [3, 3.5, 4, 4.5, 5, 5.5, 6, 6.5, 7, 7.5, 8, 9, 10, 11, 12],
-		w: [3.5, 4, 4.5, 5, 5.5, 6, 6.5, 7, 7.5, 8, 8.5, 9, 10.5, 12, 12.5]
-	},
-	eu: {m: unisex.eu, w: unisex.eu},
-	br: {m: unisex.br, w: unisex.br},
-	cm: {m: unisex.cm, w: unisex.cm},
-	in: {m: unisex.in, w: unisex.in}
-};
+import {sizes as defaultSizes} from './sizes/default.js'; // eslint-disable-line import/extensions
 
 const genders = ['w', 'm', 'k'];
 
@@ -275,6 +231,5 @@ export {
 	convertSizeRange,
 	getClosestValidSize,
 	inchesStringToNumber,
-	inchesNumberToString,
-	defaultSizes as sizes
+	inchesNumberToString
 };
